refactor(graduate): 린터 규칙 적용
