+++ apps/graduate/src/routes/_clientLayout/client.tsx
@@ -1,4 +1,5 @@
 import { createFileRoute } from '@tanstack/react-router'
+
 import { HomePage } from '~/pages/client/home'
 
 export const Route = createFileRoute('/_clientLayout/client')({
