+++ apps/graduate/src/shared/components/Section/Section.css.ts
@@ -1,4 +1,5 @@
 import { style } from '@vanilla-extract/css'
+
 import { vars } from '~/vars.css'
 
 export const section = style({
