+++ apps/graduate/src/shared/components/index.ts
@@ -1,6 +1,7 @@
-export * from './Button';
-export * from './Toolbar';
-export * from './Header';
-export * from './Pagination';
-export * from './DataTable';
-export * from './StudentAddModal';
+export * from './Button'
+export * from './Toolbar'
+export * from './Header'
+export * from './Pagination'
+export * from './DataTable'
+export * from './StudentAddModal'
+export * from './Section'
